Revert "Contact info page tests"
