+++ mobile/test/ProfileSettingsPage/ProfileSettingsPage_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.0.15 from annotations
-// in mobile/test/pages/ProfileSettingsPage/ProfileSettingsPage_test.dart.
+// in mobile/test/ProfileSettingsPage/ProfileSettingsPage_test.dart.
 // Do not manually edit this file.
 
 import 'dart:async' as _i17;
